// src/types.ts
export interface Task {
    id?: string;          // Optional for new tasks (auto-generated by Firestore)
    title: string;
    dueOn: string;        // Store dates as ISO strings (e.g., "2025-03-16")
    // taskStatus: "TO-DO" | "IN-PROGRESS" | "COMPLETED";
    // taskCategory: "WORK" | "PERSONAL";
    taskStatus: string;
    taskCategory: string;
  }